+++ src/Batch.ts
@@ -107,9 +107,7 @@ export class Batch {
   public async write(): Promise<Map<bigint, APIError>> {
     const headers: Record<string, string> = {};
     const chunks: Buffer[] = [];
-    let contentLength = 0;
     for (const [ts, { data, contentType, labels }] of this.items()) {
-      contentLength += data.length;
       chunks.push(data);
       const headerName = `x-reduct-time-${ts}`;
 
@@ -132,21 +130,11 @@ export class Batch {
     let response;
     switch (this.type) {
       case BatchType.WRITE: {
-        const stream = new ReadableStream<Uint8Array>({
-          start(ctrl) {
-            for (const chunk of chunks) {
-              ctrl.enqueue(chunk);
-            }
-            ctrl.close();
-          },
-        });
-
         headers["Content-Type"] = "application/octet-stream";
-        headers["Content-Length"] = contentLength.toString();
 
         response = await this.httpClient.post(
           `/b/${this.bucketName}/${this.entryName}/batch`,
-          stream,
+          Buffer.concat(chunks),
           headers,
         );
         break;