+++ src/Record.ts
@@ -153,8 +153,10 @@ export class WritableRecord {
 
     const headers: Record<string, string> = {
       "Content-Type": options.contentType ?? "application/octet-stream",
-      "Content-Length": contentLength.toString(),
     };
+    if (dataToSend instanceof ReadableStream) {
+      headers["x-reduct-content-length"] = contentLength.toString();
+    }
 
     for (const [key, value] of Object.entries(options.labels ?? {})) {
       headers[`x-reduct-label-${key}`] = value.toString();